Backend/Akshaya/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.15.4
inflection==0.5.1
packaging==24.0
platformdirs==4.2.2
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.26.3
